% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buildings0809}
\alias{buildings0809}
\title{Small data set of Idaho buildings}
\format{A data frame with 1248 rows and 18 variables:
\describe{
  \item{Year}{The year - 2008 or 2009}
  \item{Month}{The numeric number for month of the year}
  \item{Census.FIPS}{Federal Information Processing Standards used by the Census}
  \item{FIPS.county}{The Federal Information Processing Standards for county}
  \item{FIPS.state}{The Federal Information Processing Standards for state}
  \item{ProjectTitle}{The name of the project}
  \item{Address}{The physicall address}
  \item{City}{The city name}
  \item{State}{The state id}
  \item{Zipcode}{The 6 number zipcode}
  \item{County}{The county name}
  \item{Type}{Generalized grouping labels}
  \item{Ownership}{One of four levels - county, federal, private,   state)}
  \item{Strys}{Number of stories for the building}
  \item{SqFt}{Number of square feet built}
  \item{Value1000}{Estimated value of the building in $1,000}
  \item{Dwellings}{if apartments, how many duellings were made}
  \item{sizegroup}{a grouping of square footage size of buildings - Large,  Medium,   Small,  XLarge XXLarge}
  ...
}}
\source{
\url{https://github.com/hathawayj/buildings}
}
\usage{
buildings0809
}
\description{
A dataset containing information on buildings built in Idaho during 2008 and 2009.
}
\keyword{datasets}
