% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_zone_fips}
\alias{climate_zone_fips}
\title{County and State level FIPS numbers DOE Climate Zones}
\format{A data frame with 3221 rows and 19 variables:
\describe{
  \item{State_name}{full name of state}
  \item{State_abb}{abbreviated name of state}
  \item{FIPS.state}{Federal Information Processing Standards for each state}
  \item{AreaName}{County name with the word County}
  \item{County}{County name with the two letter state abbreviation}
  \item{Region}{Census region}
  \item{Division}{Census division}
  \item{FIPS.county}{the Federal Information Processing Standards for county}
  \item{LandSQMI}{Total area in county that is not water}
  \item{AreaSQMI}{Total area in county}
  \item{Pop100}{A census population count}
  \item{HU100}{A census housing unit count}
  \item{IntPtLat}{Internal Point (Latitude)}
  \item{IntPtLon}{Internal Point (Longitude)}
  \item{doe_climate_zone}{Department of Energy Climate Zones}
  \item{doe_moisture_regime}{The Department of Energy moisture assignment}
  \item{is_warm_humid}{The Department of Energy warm/humid assignment}
  \item{merged}{a lower case version of County column}
  \item{FIPS.long}{The long FIPS that has state/county coded}
  ...
}}
\source{
\url{https://github.com/hathawayj/buildings}
}
\usage{
climate_zone_fips
}
\description{
A dataset containing county level information for varied county aggregations.
Of most uniqueness is the department of energy climate zones as of 2010.
}
\keyword{datasets}
