% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{permits}
\alias{permits}
\title{Permit Data of Residential Buildings by county (1997 - 2008)}
\format{A data frame with 215670 rows and 8 variables:
\describe{
  \item{State}{The FIPS number for state}
  \item{Statename}{The name of the state}
  \item{county}{The FIPS number for county}
  \item{countyname}{The county name}    
  \item{variable}{The type of permit}
  \item{Year}{Year of permit data}
  \item{Month}{Month of permit data}
  \item{value}{The number of permits}
  ...
}}
\source{
\url{https://github.com/hathawayj/buildings}
}
\usage{
permits
}
\description{
A dataset containing varied types of residential permits by county.
}
\keyword{datasets}
