% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{restaurants}
\alias{restaurants}
\title{Small data set mapping restaurant names to type of restaurant}
\format{A data frame with 487 rows and 2 variables:
\describe{
  \item{Restaurant}{The name of the restaurant}
  \item{Month}{The type of dining experience - Ice Cream, Casual Dining, coffee, Fast Casual, Fast Food, Ice Cream, Pizza}
  ...
}}
\source{
\url{https://github.com/hathawayj/buildings}
}
\usage{
restaurants
}
\description{
A dataset containing information on restaurants in the US.
}
\keyword{datasets}
