% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buildings0809}
\alias{buildings0809}
\title{Small data set of Idaho buildings}
\format{A data frame with 3221 rows and 19 variables:
\describe{
  \item{Year}
  \item{Month}
  \item{Census.FIPS}
  \item{FIPS.county}{the Federal Information Processing Standards for county}
  \item{FIPS.state}{the Federal Information Processing Standards for state}
  \item{ProjectTitle}{the name of the project}
  \item{Address}{The physicall address}
  \item{City}
  \item{State}
  \item{Zipcode}
  \item{County}
  \item{Type}{Generalized grouping labels}
  \item{Ownership}{One of four levels - county, federal, private,   state)}
  \item{Strys}{Number of stories for the building}
  \item{SqFt}{Number of square feet built}
  \item{Value1000}{Estimated value of the building in $1,000}
  \item{Dwellings}{if apartments, how many duellings were made}
  \item{sizegroup}{a grouping of square footage size of buildings - Large,  Medium,   Small,  XLarge XXLarge}
  ...
}}
\source{
\url{https://github.com/hathawayj/buildings}
}
\usage{
buildings0809
}
\description{
A dataset containing information on buildings built in Idaho during 2008 and 2009.
}
\keyword{datasets}
